updated script

+++ VirtualMachineService/scripts/bash/mount.sh
@@ -1,10 +1,13 @@
 #!/bin/bash
 sudo touch test
-declare -a volumes=VOLUME_IDS
-declare -a paths=VOLUME_PATHS
+declare -a volumes_new=VOLUME_IDS_NEW
+declare -a paths_new=VOLUME_PATHS_NEW
+declare -a volumes_attach=VOLUME_IDS_ATTACH
+declare -a paths_attach=VOLUME_PATHS_ATTACH
+
 ITER=0
 ABORT_COUNTER=0
-for id in "${volumes[@]}"; do
+for id in "${volumes_new[@]}"; do
   while :; do
     if [ -e /dev/disk/by-id/"$id" ]; then
       echo "volume  ""$id"" found"
@@ -13,17 +16,53 @@ for id in "${volumes[@]}"; do
     else
       echo "no volume ""$id"" found"
       if [ $ABORT_COUNTER -gt 12 ]; then
-        exit 1
+        break
       fi
       ((ABORT_COUNTER++))
 
       sleep 10
     fi
   done
+  if [ $ABORT_COUNTER -gt 12 ]; then
+    echo "Waited 120 seconds for ""$id"" "
+    ABORT_COUNTER=0
+
+    continue
+  fi
   cd /dev/disk/by-id || exit
   sudo mkfs.ext4 "$id"
-  sudo mkdir -p /mnt/${paths[ITER]}
-  sudo chmod 777 /mnt/${paths[ITER]}/
-  sudo mount "$id" /mnt/${paths[ITER]}
+  sudo mkdir -p /mnt/${paths_new[ITER]}
+  sudo chmod 777 /mnt/${paths_new[ITER]}/
+  sudo mount "$id" /mnt/${paths_new[ITER]}
+  ((ITER++))
+done
+ITER=0
+ABORT_COUNTER=0
+for id in "${volumes_attach[@]}"; do
+  while :; do
+    if [ -e /dev/disk/by-id/"$id" ]; then
+      echo "volume  ""$id"" found"
+      ABORT_COUNTER=0
+      break
+    else
+      echo "no volume ""$id"" found"
+      if [ $ABORT_COUNTER -gt 12 ]; then
+        break
+      fi
+      ((ABORT_COUNTER++))
+
+      sleep 10
+    fi
+  done
+  if [ $ABORT_COUNTER -gt 12 ]; then
+    echo "Waited 120 seconds for ""$id"" "
+    ABORT_COUNTER=0
+
+    continue
+  fi
+  cd /dev/disk/by-id || exit
+  sudo mkdir -p /mnt/${paths_attach[ITER]}
+  sudo chmod 777 /mnt/${paths_attach[ITER]}/
+  sudo mount "$id" /mnt/${paths_attach[ITER]}
   ((ITER++))
 done
